LPS-58897 Move out common test classes to bundle
