+++ modules/portal-cache/portal-cache-test-common/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Portal Cache Test Common
+Bundle-SymbolicName: com.liferay.portal.cache.test.common
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.cache.test.common
+Include-Resource: classes
