+++ modules/portal-cache/portal-cache-test-common/src/com/liferay/portal/cache/test/common/TestPortalCacheManager.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.cache.test;
+package com.liferay.portal.cache.test.common;
 
-import com.liferay.portal.kernel.cache.AbstractPortalCacheManager;
+import com.liferay.portal.cache.AbstractPortalCacheManager;
+import com.liferay.portal.cache.configuration.PortalCacheConfiguration;
+import com.liferay.portal.cache.configuration.PortalCacheManagerConfiguration;
 import com.liferay.portal.kernel.cache.PortalCache;
-import com.liferay.portal.kernel.cache.configuration.PortalCacheConfiguration;
-import com.liferay.portal.kernel.cache.configuration.PortalCacheManagerConfiguration;
 
 import java.io.Serializable;
