+++ modules/portal-cache/portal-cache-test-common/src/com/liferay/portal/cache/test/common/TestPortalCacheReplicator.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.cache.test;
+package com.liferay.portal.cache.test.common;
 
-import com.liferay.portal.kernel.cache.PortalCacheReplicator;
+import com.liferay.portal.cache.PortalCacheReplicator;
 
 import java.io.Serializable;
